style: Separate files of different types
Now the interface is divided by base_windows which is  python code generated by pyside6, qt_files, which is file .ui generated by qt designer and "our creation" files
