+++ interface/ui-qt/base_windows/delete_confirmation.py
