+++ interface/ui-qt/base_windows/register_company.py
