+++ interface/ui-qt/base_windows/register_person.py
