+++ interface/ui-qt/base_windows/sucessful_register.py
