+++ interface/ui-qt/images/sucessful_register.ui
