+++ interface/ui-qt/qt_files/delete_confirmation.ui
