+++ interface/ui-qt/qt_files/register_person.ui
